Split out GVStyle from greyVest index
